requirements.txt
Django==3.2.20
djangorestframework==3.14.0
